* implemented new mission category: intercept. UFO can fly around the earth to look for PHALANX aircraft to destroy. * use aircraft_t->mission to store the mission assigned to a UFO -- BREAKS SAVEGAME * removed console commands: 'addufo' (you should use debug_addmission instead) and 'debug_destroyallufos'

+++ src/client/cl_basemanagement.h
@@ -225,7 +225,7 @@ typedef struct base_s {
 
 /* @todo: remove isDiscovered when alienInterest is implemented */
 	qboolean isDiscovered;	/** qtrue if the base has been discovered by aliens */
-	int alienInterest;		/**< How much aliens know this base (and may attack it) */
+	float alienInterest;		/**< How much aliens know this base (and may attack it) */
 
 	radar_t	radar;	/**< the onconstruct value of the buliding building_radar increases the sensor width */
